multiple source files

+++ scripts/generate-server.js
@@ -1,23 +1,23 @@
 import * as path from 'path'
 import * as fs from 'fs'
-import { getRouteStubs, sourceFilePath } from './shared/route-stubs.js'
+import { getRouteStubs } from './shared/route-stubs.js'
+import { getGeneratorConfig } from './shared/generator-config.js'
+import { getAppRootPath } from './shared/app-path.js'
 
 const routesFileTemplatePath = path.resolve('templates/express-routes-file.template.txt')
 const routesFileTemplate = fs.readFileSync(routesFileTemplatePath, 'utf8')
 const generatedPackageTemplatePath = path.resolve('templates/generated-package.template.json')
 const generatedPackageTemplate = fs.readFileSync(generatedPackageTemplatePath, 'utf8')
-const generatedRunnerTemplatePath = path.resolve('templates/generated-runner.template.txt')
-const generatedRunnerTemplate = fs.readFileSync(generatedRunnerTemplatePath, 'utf8')
 const generatedTsConfigTemplatePath = path.resolve('templates/generated-tsconfig.template.json')
 const generatedTsConfigTemplate = fs.readFileSync(generatedTsConfigTemplatePath, 'utf8')
 
-const routeStubs = getRouteStubs()
+const appRootPath = getAppRootPath()
+const routeStubsByKey = getRouteStubs(appRootPath)
+const generatorConfig = getGeneratorConfig(appRootPath)
 
 function createExpressRoutesFile() {
-    const destDir = path.resolve('../generated/server')
+    const destDir = path.join(appRootPath, 'generated/server')
     const destFilePath = path.join(destDir, 'express-routes.generated.ts')
-    const copiedSourcePath = path.join(destDir, path.basename(sourceFilePath))
-    const runnerFilePath = path.join(destDir, 'run-generated.ts')
     const packageJsonPath = path.join(destDir, 'package.json')
     const tsConfigPath = path.join(destDir, 'tsconfig.json')
 
@@ -25,27 +25,51 @@ function createExpressRoutesFile() {
         fs.mkdirSync(destDir, { recursive: true })
     }
 
-    const methodNames = routeStubs.map((stub) => stub.methodName)
-    const importsLine = methodNames.length > 0
-        ? `import { ${methodNames.join(', ')} } from './index-server'`
-        : ''
+    const targets = generatorConfig.targets ?? ['index']
+    const copiedSourceFiles = []
+    const importsLines = []
 
-    const routesCode = routeStubs.map((stub) => stub.routeBody).join('\n\n')
+    for (const key of targets) {
+        const serverFileName = `${key}-server.ts`
+        const sourceFilePath = path.join(appRootPath, serverFileName)
+        const copiedSourcePath = path.join(destDir, serverFileName)
+
+        if (!fs.existsSync(sourceFilePath)) {
+            throw new Error(`Could not find ${serverFileName}`)
+        }
+
+        fs.copyFileSync(sourceFilePath, copiedSourcePath)
+        copiedSourceFiles.push(copiedSourcePath)
+
+        const methodNames = [...new Set((routeStubsByKey[key] ?? []).map((stub) => stub.methodName))]
+
+        if (methodNames.length > 0) {
+            importsLines.push(`import { ${methodNames.join(', ')} } from './${serverFileName.replace('.ts', '')}'`)
+        }
+    }
+
+    const routesBodies = []
+    for (const [key, stubs] of Object.entries(routeStubsByKey)) {
+        for (const stub of stubs) {
+            routesBodies.push(stub.routeBody)
+        }
+    }
+
+    const importsBlock = importsLines.join('\n')
+    const routesCode = routesBodies.join('\n\n')
 
     const fileContent = routesFileTemplate
-        .replaceAll('{{importsLine}}', importsLine)
+        .replaceAll('{{corsOptions}}', JSON.stringify(generatorConfig.cors ?? { origin: '*' }))
+        .replaceAll('{{importsLine}}', importsBlock)
         .replaceAll('{{routesCode}}', routesCode)
 
     fs.writeFileSync(destFilePath, fileContent, 'utf8')
-    fs.copyFileSync(sourceFilePath, copiedSourcePath)
-    fs.writeFileSync(runnerFilePath, generatedRunnerTemplate, 'utf8')
     fs.writeFileSync(packageJsonPath, generatedPackageTemplate, 'utf8')
     fs.writeFileSync(tsConfigPath, generatedTsConfigTemplate, 'utf8')
 
     return {
         routesFile: destFilePath,
-        copiedSourceFile: copiedSourcePath,
-        runnerFile: runnerFilePath,
+        copiedSourceFiles,
         packageJsonFile: packageJsonPath,
         tsConfigFile: tsConfigPath
     }
@@ -53,4 +77,4 @@ function createExpressRoutesFile() {
 
 createExpressRoutesFile()
 
-console.log(JSON.stringify(routeStubs, null, 2))
+console.log(JSON.stringify(routeStubsByKey, null, 2))
